Moving context hooks to hooks directory
