+++ src/hooks/useElementInView.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { ViewPortContext } from '~/providers/ViewPortProvider';
+
+const useElementInView = () => useContext(ViewPortContext);
+
+export default useElementInView;
