+++ src/hooks/useNotifications.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { NotificationContext } from '~/providers/NotificationProvider';
+
+const useNotifications = () => useContext(NotificationContext);
+
+export default useNotifications;
