+++ src/hooks/useTheme.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { ThemeContext } from '~/providers/ThemeProvider';
+
+const useTheme = () => useContext(ThemeContext);
+
+export default useTheme;
