+++ src/hooks/useThreeAnimation.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { ThreeAnimationContext } from '~/providers/ThreeAnimationProvider';
+
+const useThreeAnimation = () => useContext(ThreeAnimationContext);
+
+export default useThreeAnimation;
